% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{BiocBPRPContainer-class}
\alias{BiocBPRPContainer-class}
\alias{.BiocBPRPContainer}
\title{The Bioconductor BiocParallel Redis container}
\description{
The Bioconductor BiocParallel Redis container. See \code{?BiocBPRPServerContainer} and
\code{?BiocBPRPWorkerContainer}
}

