% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BiocBPRPContainer.R
\name{BiocBPRPContainer-commom-parameters}
\alias{BiocBPRPContainer-commom-parameters}
\title{Common BiocBPRPContainer parameter}
\arguments{
\item{image}{Character, the container image}

\item{name}{Character, the optional name of the container}

\item{environment}{List, the environment variables in the container}
}
\description{
Common BiocBPRPContainer parameter
}
