% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BiocBPRPContainer.R
\name{BiocBPRPServerContainer}
\alias{BiocBPRPServerContainer}
\title{Get the Bioconductor BiocParallel RedisParam container}
\usage{
BiocBPRPServerContainer(environment = list(), tag = "latest")
}
\arguments{
\item{environment}{List, the environment variables in the container}
}
\value{
a \code{BiocBPRPContainer} object
}
\description{
Get the Bioconductor BiocParallel RedisParam container.
}
\examples{
BiocBPRPServerContainer()
}
