% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BiocBPRPContainer.R
\name{BiocBPRPWorkerContainer}
\alias{BiocBPRPWorkerContainer}
\title{Get the Bioconductor BiocParallel Redis worker container}
\usage{
BiocBPRPWorkerContainer(
  RPackages = NULL,
  sysPackages = NULL,
  environment = list(),
  maxWorkerNum = 4L,
  tag = "latest"
)
}
\arguments{
\item{RPackages}{Character, a vector of R packages that will be installed
by \code{AnVIL::install} before connecting with the server}

\item{sysPackages}{Character, a vector of system packages that will be installed
by \verb{apt-get install} before running the R worker}

\item{environment}{List, the environment variables in the container}

\item{maxWorkerNum}{Integer, the maximum worker number in a container}
}
\value{
a \code{BiocBPRPContainer} object
}
\description{
Get the Bioconductor BiocParallel Redis worker container.
}
\examples{
BiocBPRPWorkerContainer()
}
