% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-BiocBPRPContainer.R
\name{getServerContainer,BiocBPRPContainer-method}
\alias{getServerContainer,BiocBPRPContainer-method}
\title{Get the Bioconductor BiocParallel Redis container}
\usage{
\S4method{getServerContainer}{BiocBPRPContainer}(workerContainer)
}
\arguments{
\item{workerContainer}{The worker container.}
}
\value{
A \code{BiocBPRPContainer} server container
}
\description{
Get the Bioconductor BiocParallel Redis container from the worker container
}
